Avoid NVRAM out-of-space issue (#1276)
PR 1120 added a TPM NVRAM behavior where the AKCert NVRAM index is
always undefined and redefined, with size 4kB and the platform-defined
attribute. In some cases, existing VMs may have an NVRAM blob that is
too small to accommodate a 4kB index. This change works around that by
not modifying the existing index if it is not platform-defined. This
change also keeps other behavior added in 1120, such as preserving any
existing contents of that NVRAM index.