uh_mem: Properly reset permissions when unsharing pages (#1891) (#1894)
When we are sharing a page we remove all VTL 0 permissions to that page.
Later on, when we re-private the page, we were failing to reset these
permissions, which led to failures when the guest tried to use pages it
should have had access to. This code is a bit confusing due to
conflating private/shared with VTL 1 access permissions. Add a bunch of
comments, and fix the reset.

Cherry-pick of #1891