netvsp: handle RNDIS packet filter OID for stopping Rx (#1967)
Linux netvsc sends an OID to stop receiving packets on vmbus channel
close. Example scenarios: hibernation and MTU change. Prior to opening a
new channel and processing the packets, netvsc checks that there are no
pending packets. If there are, netvsc logs and error and is unable to
recover. We observe the error: `hv_netvsc eth0: Ring buffer not empty
after closing rndis` in the guest syslog.

Modifying netvsp to handle the OID and drop incoming RX traffic. This
will allow for netvsc to successfully close and re-open the vmbus
channel, even under heavy incoming traffic.

Cherry-pick of #1873 and #1949

---------

Co-authored-by: erfrimod <31358361+erfrimod@users.noreply.github.com>
Co-authored-by: Sunil Muthuswamy <sunilmut@microsoft.com>
Co-authored-by: Brian Perkins <brian.perkins@microsoft.com>