Cleanup intercept device shutdown (#2140)
1. Tie lifetime of the device to the vmbus parent a. Remove existing
channel used to control lifetime.
2. If the vmbus relay parent is dropped, there will be no opportunity to
close the GPADL associated with the special ring buffer memory pages. In
this case, note the error and leak the device. Trying to change the VTL
permissions on these memory pages while the host still has them open
will result in an error.
3. Allow intercepted devices to be revoked and reoffered from the host.

CP of #2117

Co-authored-by: Brian Perkins <brian.perkins@microsoft.com>