petri/pipette: extend timeout for connect to workaround windows vmbus bug (#2485) (#2489)
On CVMs, there's a Windows guest bug where a revoked hvsock with a
listener inside the guest will cause a crash. Until this bug is fixed in
public builds, attempt to workaround the issue by extending the
connection time to allow the guest to boot and the agent to start
listening.

If this still doesn't work, we may need to try a slightly different
strategy. See #2470.

Clean cherry pick of #2485