Renew AKCert if platform-defined and vTPM blob is 32k (#2316) (#2497)
HCL-provisioned vTPMs have a platform-defined AKCert index and the
AKCert has to be renewed by OpenHCL. We were planning on using control
flags (see, e.g., PR #2036) to control whether OpenHCL will renew the
AKCert, but it needs to run on older host OS versions that don't support
these flags.

As an alternative, we can detect that the existing vTPM has a
platform-defined AKCert index and is 32kB in size, which indicates that
the vTPM was provisioned by HCL and that OpenHCL should handle renewing
the AKCert.

This change also adds handling for a power-fail-robustness case, where
the VM comes up without an AKCert index at all. In that case, HCL should
try to create it if the vTPM is at least 32kB.