openhcl_boot: include more info in the panic message when unable to allocate private pool (#2590) (#2601)
Clean cherry pick of #2590.

We have observed some tests fail when allocating the private pool. Those
tests do not enable COM logging, so we are unable to ascertain the
parameters leading to why the memory size was chosen. This panic is
included in the guest crash message, so add some extra info there.