emuplat/nvme_manager: add pci_id to logs when failing to create/restore a vfio device (#2543) (#2599)
Clean cherry pick of #2543.

I have seen two test failures where OpenHCL cannot restore because the
vfio restore path cannot find the iommu group file.
From those logs, I think this is just a case where the kernel is async
and is yet to create the path. However, I also
identified that it is not *crystal clear* which device triggered the
failure. So, add those logs.